Time: 27 ms (55.48%), Space: 89.6 MB (81.72%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.cpp
@@ -21,7 +21,7 @@ double findMedianSortedArrays(vector<int>& A, vector<int>& B) {
             if((n1+n2)%2!=0){
                 return min(Aright, Bright);
             }else{
-                return ((long double)max(Aleft,Bleft)+min(Aright,Bright))/2;
+                return (max(Aleft,Bleft)+min(Aright,Bright))/2.0;
             }
         }
